LPS-88338 Add new methods to HttpUtil instead of PortalUtil
