+++ portal-kernel/src/com/liferay/portal/kernel/util/Http.java
@@ -106,6 +106,8 @@ public interface Http {
 
 	public String getProtocol(String url);
 
+	public String getQueryString(HttpServletRequest httpServletRequest);
+
 	public String getQueryString(String url);
 
 	public String getRequestURL(HttpServletRequest httpServletRequest);
@@ -117,6 +119,8 @@ public interface Http {
 	public boolean hasProtocol(String url);
 
 	public boolean hasProxyConfig();
+
+	public boolean isForwarded(HttpServletRequest httpServletRequest);
 
 	public boolean isNonProxyHost(String host);
