+++ portal-kernel/src/com/liferay/portal/kernel/util/Portal.java
@@ -958,8 +958,6 @@ public interface Portal {
 
 	public PreferencesValidator getPreferencesValidator(Portlet portlet);
 
-	public String getQueryString(HttpServletRequest httpServletRequest);
-
 	public String getRelativeHomeURL(HttpServletRequest httpServletRequest)
 		throws PortalException;
 
@@ -1154,8 +1152,6 @@ public interface Portal {
 		String portletId, ThemeDisplay themeDisplay);
 
 	public boolean isCustomPortletMode(PortletMode portletMode);
-
-	public boolean isForwarded(HttpServletRequest httpServletRequest);
 
 	public boolean isForwardedSecure(HttpServletRequest httpServletRequest);
